chore: update the project structure to support angular 16
changed the files and structure of the project to be more similar to current angular standards, this
will enable ivy engine and thus it will work on angular 16

BREAKING CHANGE: Requires angular v16 and rxjs v7

fix #59
