+++ demo/app/child/child.component.ts
@@ -1,9 +1,11 @@
 import { Component, ChangeDetectionStrategy } from '@angular/core';
 import { TimeagoIntl } from 'ngx-timeago';
-import { strings as englishStrings } from 'ngx-timeago/language-strings/en';
-import { strings as englishShortStrings } from 'ngx-timeago/language-strings/en-short';
-import { strings as frenchStrings } from 'ngx-timeago/language-strings/fr';
-import { strings as frenchShortStrings } from 'ngx-timeago/language-strings/fr-short';
+
+// These imports are wrong, we need to find a way to correctly export the language-string
+import { strings as englishStrings } from '../../../lib/src/language-strings/en';
+import { strings as englishShortStrings } from '../../../lib/src/language-strings/en-short';
+import { strings as frenchStrings } from '../../../lib/src/language-strings/fr';
+import { strings as frenchShortStrings } from '../../../lib/src/language-strings/fr-short';
 
 @Component({
   selector: 'app-child',
