--- demo/tsconfig.packages.json
@@ -1,23 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "angularCompilerOptions": {
-    "paths": {
-      "ngx-timeago": [ "../dist/ngx-timeago" ],
-      "ngx-timeago/*": [ "../dist/ngx-timeago/*" ]
-    }
-  },
-  "compilerOptions": {
-    "outDir": "../out-tsc/app",
-    "module": "es2015",
-    "baseUrl": "",
-    "types": [],
-    "paths": {
-      "ngx-timeago": [ "../dist/ngx-timeago" ],
-      "ngx-timeago/*": [ "../dist/ngx-timeago/*" ]
-    }
-  },
-  "exclude": [
-    "test.ts",
-    "**/*.spec.ts"
-  ]
-}
