--- demo/tsconfig.spec.json
@@ -1,29 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../out-tsc/spec",
-    "module": "commonjs",
-    "target": "es5",
-    "baseUrl": "",
-    "types": [
-      "jasmine",
-      "node"
-    ],
-    "paths": {
-      "ngx-timeago": [ "../dist/ngx-timeago" ],
-      "ngx-timeago/*": [ "../dist/ngx-timeago/*" ]
-    }
-  },
-  "files": [
-    "test.ts",
-    "polyfills.ts"
-  ],
-  "include": [
-    "../tests/**/*.spec.ts",
-    "../tests/**/*.d.ts"
-  ],
-  "exclude": [
-    "node_modules",
-    "tmp"
-  ]
-}
