--- demo/typings.d.ts
@@ -1,5 +1,0 @@
-/* SystemJS module definition */
-declare var module: NodeModule;
-interface NodeModule {
-  id: string;
-}
