+++ lib/tsconfig.spec.json
@@ -1,13 +1,13 @@
 {
   "extends": "../tsconfig.json",
   "compilerOptions": {
-    "outDir": "../out-tsc/app",
-    "types": []
+    "outDir": "../out-tsc/spec",
+    "types": [
+      "jasmine"
+    ]
   },
-  "files": [
-    "main.ts"
-  ],
   "include": [
+    "**/*.spec.ts",
     "**/*.d.ts"
   ]
 }
