+++ package.json
@@ -3,63 +3,46 @@
   "private": true,
   "license": "MIT",
   "scripts": {
-    "ng:build:demo": "ng build packages --prod --base-href=https://ihym.github.io/ngx-timeago/",
-    "build:demo": "cpr dist/ngx-timeago demo/ngx-timeago && npm run ng:build:demo && rimraf demo/ngx-timeago",
-    "build:lib": "rimraf dist && ng-packagr -p lib/package.json && gulp",
-    "lint": "ng lint",
+    "build:demo": "ng build ngx-timeago-demo --base-href=https://ihym.github.io/ngx-timeago/",
+    "build:lib": "ng build ngx-timeago",
     "ng": "ng",
     "start": "ng serve",
-    "test": "ng test packages --watch=false",
-    "tdd": "ng test dev --watch",
+    "test": "ng test",
+    "tdd": "ng test --watch",
     "commit": "git-cz",
-    "analyze": "npm run build:lib && cpr dist/ngx-timeago demo/ngx-timeago && npm run ng:build:demo  -- --stats-json && rimraf demo/ngx-timeago && webpack-bundle-analyzer dist/demo/stats.json"
+    "analyze": "npm run build:lib && npm run build:demo  -- --stats-json && webpack-bundle-analyzer dist/demo/stats.json"
   },
   "dependencies": {
-    "@angular/animations": "10.0.8",
-    "@angular/cdk": "10.1.2",
-    "@angular/common": "10.0.8",
-    "@angular/compiler": "10.0.8",
-    "@angular/core": "10.0.8",
-    "@angular/forms": "10.0.8",
-    "@angular/material": "10.1.2",
-    "@angular/platform-browser": "10.0.8",
-    "@angular/platform-browser-dynamic": "10.0.8",
-    "@angular/router": "10.0.8",
-    "core-js": "3.6.5",
-    "rxjs": "6.6.2",
-    "zone.js": "0.10.3"
+    "@angular/animations": "^16.0.0",
+    "@angular/cdk": "^16.0.1",
+    "@angular/common": "^16.0.0",
+    "@angular/compiler": "^16.0.0",
+    "@angular/core": "^16.0.0",
+    "@angular/forms": "^16.0.0",
+    "@angular/material": "^16.0.1",
+    "@angular/platform-browser": "^16.0.0",
+    "@angular/platform-browser-dynamic": "^16.0.0",
+    "@angular/router": "^16.0.0",
+    "bootstrap": "4.5.2",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.13.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.1000.5",
-    "@angular/cli": "10.0.5",
-    "@angular/compiler-cli": "10.0.8",
-    "@angular/language-service": "10.0.8",
-    "@types/jasmine": "3.5.12",
-    "@types/node": "14.0.27",
-    "bootstrap": "4.5.2",
-    "codelyzer": "6.0.0",
+    "@angular-devkit/build-angular": "^16.0.1",
+    "@angular/cli": "~16.0.1",
+    "@angular/compiler-cli": "^16.0.0",
+    "@types/jasmine": "~4.3.0",
     "commitizen": "4.1.2",
-    "cpr": "^3.0.0",
     "cz-conventional-changelog": "3.2.0",
-    "gulp": "4.0.2",
-    "gulp-typescript": "6.0.0-alpha.1",
-    "jasmine-core": "3.6.0",
-    "jasmine-spec-reporter": "5.0.2",
-    "karma": "5.1.1",
-    "karma-chrome-launcher": "3.1.0",
-    "karma-cli": "2.0.0",
-    "karma-coverage-istanbul-reporter": "3.0.3",
-    "karma-jasmine": "3.3.1",
-    "karma-jasmine-html-reporter": "1.5.4",
-    "merge2": "1.4.1",
-    "ng-packagr": "10.0.3",
-    "protractor": "7.0.0",
-    "rimraf": "3.0.2",
-    "ts-node": "8.10.2",
-    "tsickle": "0.39.1",
-    "tslib": "2.0.1",
-    "tslint": "6.1.3",
-    "typescript": "3.9.7",
+    "jasmine-core": "~4.6.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.0.0",
+    "ng-packagr": "^16.0.0",
+    "typescript": "~5.0.2",
     "webpack-bundle-analyzer": "3.8.0"
   },
   "config": {
